move tests out of src for cleaner Dockerfile
