--- src/tests/integration/home.test.js
@@ -1,11 +1,0 @@
-const request = require('supertest');
-const app = require('../../startup/app');
-
-describe('home', () => {
-  it('should return the home page', async () => {
-    const res = await request(app).get('/');
-
-    expect(res.status).toBe(200);
-    expect(res.text).toMatch(/[Hh]ome/);
-  });
-});
